Remove testing backend and frontend containers from deployment workflows
The exact same commits have already been tested in the backend and
frontend repos, so running them also here just creates a possible
failure point when the test execution command has changed in one of the
repos, and it's forgotten to update to the monorepo's circleci config.
That will result in an awkward error when the CI fails only
when merging develop to master (because we didn't run the tests when
just merging to develop, to save some deployment time/resources),
and then to fix it we have to make a totally new PR to develop,
and then from develop to master.

This exact scenario happened in e08f797.