Remove `depends_on: backend` from frontend compose service
This was problematic, since if one wanted to run the backend service
with PyCharm's debugger and then just the frontend with `docker-compose
up frontend`, starting the frontend would just stop the backend
debugging session and start backend normally as a dependency.
This could've of course be circumvented by passing `--no-deps` to
compose, but it would just create an unnecessary inconvenience.