Fix edge case bug in check_submodules.sh
Previously the script would fail with this kind of graph, thinking that
`a54d444c` was only part of `dangling-feature` and not at all `develop`.
```
*   93daa1d9 <-- develop
|\
| * 09f69c5d
|/
| *   1c5f2f70 <-- dangling-feature
| |\
| |/
|/|
* |   e4b3e4a9
|\ \
| * | c956f00f
|/ /
| * 00dc6331
|/
*   a54d444c <-- commit to check
|\
| * d3a33875
| * 3aff3447
```

Also make the script loops over all submodules instead of a hard codes
list (where we had forgot to add `infra` to).

Not using `git submodule foreach --quiet` because it takes its input as
a string which is cumbersome.